+++ lib/read/cardano-wallet-read.cabal
@@ -56,6 +56,7 @@ library
     Cardano.Read.Ledger.Tx.ExtraSigs
     Cardano.Read.Ledger.Tx.Fee
     Cardano.Read.Ledger.Tx.Hash
+    Cardano.Read.Ledger.Tx.Integrity
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -88,7 +89,6 @@ library
     Cardano.Wallet.Read.Tx.Gen.Shelley
     Cardano.Wallet.Read.Tx.Gen.TxParameters
     Cardano.Wallet.Read.Tx.Inputs
-    Cardano.Wallet.Read.Tx.Integrity
     Cardano.Wallet.Read.Tx.Metadata
     Cardano.Wallet.Read.Tx.Mint
     Cardano.Wallet.Read.Tx.ReferenceInputs
